organize: Trade Methodes in TradeYMLManager
